refactor(nms): move fbc lib platform_server
