+++ nms/scripts/enableAllNetworkIDsForOrganization.js
@@ -17,8 +17,7 @@
 
 import Sequelize from 'sequelize';
 import https from 'https';
-
-import {API_HOST, apiCredentials} from '../fbc_js_core/platform_server/config';
+import {API_HOST, apiCredentials} from '../config/config';
 import {Organization} from '../fbc_js_core/sequelize_models';
 
 async function enableNetworks(
