+++ nms/server/host/middleware.js
@@ -15,9 +15,9 @@
  */
 
 import type {ExpressResponse, NextFunction} from 'express';
-import type {FBCNMSRequest} from '../../../server/auth/access';
+import type {FBCNMSRequest} from '../auth/access';
 
-import asyncHandler from '../../../fbc_js_core/util/asyncHandler';
+import asyncHandler from '../../fbc_js_core/util/asyncHandler';
 
 export const hostOrgMiddleware = asyncHandler(
   async (req: FBCNMSRequest, res: ExpressResponse, next: NextFunction) => {
