+++ nms/server/main/routes.js
@@ -23,16 +23,18 @@ import adminRoutes from '../admin/routes';
 import apiControllerRoutes from '../apicontroller/routes';
 import asyncHandler from '../../fbc_js_core/util/asyncHandler';
 import express from 'express';
+import hostRoutes from '../host/routes';
+import loggerRoutes from '../logger/routes';
 import networkRoutes from '../network/routes';
 import path from 'path';
 import staticDist from '../../fbc_js_core/webpack_config/staticDist';
+import testRoutes from '../test/routes';
 import userMiddleware from '../auth/express';
 import {AccessRoles} from '../../shared/roles';
-
 import {TABS} from '../../fbc_js_core/types/tabs';
 import {access} from '../auth/access';
-import {getEnabledFeatures} from '../../fbc_js_core/platform_server/features';
-import {hostOrgMiddleware} from '../../fbc_js_core/platform_server/host/middleware';
+import {getEnabledFeatures} from '../features';
+import {hostOrgMiddleware} from '../host/middleware';
 
 const router: express.Router<FBCNMSRequest, ExpressResponse> = express.Router();
 
@@ -80,11 +82,8 @@ router.use('/nms/apicontroller', apiControllerRoutes);
 router.use('/nms/network', networkRoutes);
 router.use('/nms/static', express.static(path.join(__dirname, '../static')));
 
-router.use(
-  '/logger',
-  require('../../fbc_js_core/platform_server/logger/routes'),
-);
-router.use('/test', require('../../fbc_js_core/platform_server/test/routes'));
+router.use('/logger', loggerRoutes);
+router.use('/test', testRoutes);
 router.use(
   '/user',
   userMiddleware({
@@ -102,8 +101,7 @@ router.get(
   }),
 );
 
-const hostRouter = require('../../fbc_js_core/platform_server/host/routes');
-router.use('/host', hostOrgMiddleware, hostRouter.default);
+router.use('/host', hostOrgMiddleware, hostRoutes);
 
 async function handleHost(req: FBCNMSRequest, res) {
   const appData: AppContextAppData = {
